requirements.txt
flask
flask-sqlalchemy
flask-login
azure-ai-textanalytics